+++ src/components/molecules/DragonBank/styles/index.scss
@@ -1,2 +1,7 @@
 .dragon-bank {
+  display: flex;
+  flex-flow: row nowrap;
+  align-items: center;
+  justify-content: center;
+  gap: 2.4rem;
 }
