--- src/components/molecules/DropZone/styles/index.scss
@@ -1,5 +1,0 @@
-.drop-zone {
-  span:not(.sticker) {
-    display: none !important;
-  }
-}
