+++ src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 export * from './Countdown';
-export {} from './Submission';
+export * from './DragonBank';
+export * from './Submission';
 export * from './Table';
